CC mode, 4400mv profile

+++ drivers/power/tusb422_linux_v1.00/usb_pd_policy_engine.h
@@ -311,8 +311,8 @@ typedef struct
 	uint8_t             object_position;  /* Range: 1 - 7 */
 	uint8_t             selected_snk_pdo_idx;
 	uint8_t		    apdo_idx;	/* if src has PPS support, it tracks PPS PDO index */
-	uint32_t	    apdo_max_volt;
-	uint32_t	    apdo_max_curr;
+	uint16_t	    apdo_max_volt;
+	uint16_t	    apdo_max_curr;
 
 	uint32_t            rdo;
 	uint32_t            selected_pdo;
